rect select + optimization
added rectangle selection for left click + drag.
currently, the selection does nothing (not used by anything and no visual effect to know if something is rendered) - this will be added in the next update for graphics.

optimized the code used by this and drag action to not perform redundant frame updates.

added helper functions for QRectF creation/manipulation.