improved select
added a bool drag_moved that flags if the mouse moved (at all ever) during a drag event.

added a select function for handling a single node.
if at the end of a drag event drag_moved is false, the node is selected via this function.

changed the "add" logic of the selection functions to "mod".
no-mod (not holding control) is still the same (but not optimized for speed).
mod form now toggles items into or out of the selection instead of just unioning (effectively xor select).

moved all the settings into the implementation file (makes more sense that way anyway).