added background context menu
added a context menu that pops up when you right click on the background.
currently, all it has is Add, which adds a new node at the cursor position.
because this invalidates iterators, it also clears the current selection.
this could be changed to preserve the current selection at the expense of speed (mainly a problem if AdventureMap later on uses something other than a random-access container), but i don't really see why the user would care either way.