added arc editor
implemented the arc editor section of the node editor.
a button at the bottom adds a new arc to fill in.
to delete an arc, uncheck the "enabled" checkbox to the left of an arc.
this serves two purposes:
one: we don't have to remove widgets upon "deletion" of an arc this way.
two: this keeps the arc data around in case deletion was a misclick and avoids needing an undo system or tons of confirmation popups.